chore: DTO 적용 중간커밋
