+++ src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './UserService';
+export * from './CandiesService';
